requirements.txt
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.19.0
Faker>=13.0.0
labelme>=5.3.0
pyyaml
pyqt5
pdf2image
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
gender-guesser>=0.4.0
zipcodes>=1.1.2
openpyxl>=3.1.0